requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
click==8.1.3
Django==4.1.6
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
sqlparse==0.4.3
Werkzeug==2.2.2
zipp==3.12.0
djangorestframework~=3.14.0